requirements.txt
fastapi[standard]==0.114.2
sqlmodel==0.0.22
psycopg2-binary==2.9.9
pyjwt==2.9.0
bcrypt==4.2.0
passlib==1.7.4
pytest==8.3.3
